Make sure isExcluded() never returns None
The function should always return True or False, never None. The indentation of
the "return True" was wrong, making it inside an if. This patch just moves it
outside, making it a "catch all" return, and then fixing when the function ended
without running a return statement.

This issue broke the plugin as None evaluates as False causing the plugin to
never search for subtitles.