Change in language string
Only selected_language that was on strings.xml direct in the option.
Drop of Portuguese (Brazil) option.
There is no ISO for that language, only Portuguese is known.
So many subtitle services download the subtitles with .pt.srt in the end
or without nothing. So brazilian and portuguese subtitles are always
shown as pt. in case of brazilian sometimes arent presented as correct
language.